requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pytest==7.4.0
pyserial==3.5
setuptools==59.6.0
